adjust hard coded max_ratio_in_out_unit_flow
the value ensures the function to switch on or off a unit and at the same time do not have to much power input when producing. Currently, we do not have another way to ensure a minimal power input when not producing but are in stand-by mode and have units_on cost.